refactor: change directory structure
- add `lib` directory to store utilities
- add `components/timer` directory to store files regarding timer
  component
- fix the extension of `useTimer`
