+++ src/lib/math.test.ts
