+++ src/lib/math.ts
